Divide Trailblazer::Railtie in 2 modules: code to load and code to include
In this way in the future gem we are able to include only the "include" module so we don't load stuff twice
